+++ instrumentation/openai-client-instrumentation/instrumentation-1.1/src/test/java/co/elastic/otel/openai/v1_1/ResponseHeaderScrubber.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.otel.openai;
+package co.elastic.otel.openai.v1_1;
 
 import com.github.tomakehurst.wiremock.common.FileSource;
 import com.github.tomakehurst.wiremock.extension.Parameters;
